requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-crispy-forms==1.7.2
django-heroku==0.3.1
gunicorn==19.9.0
Pillow==6.2.0
psycopg2==2.8.3
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
